+++ inst/create_tables.php
@@ -32,6 +32,8 @@ $res = $DBH->exec($tables['Categories_2'][$config['DB']['type']]);              
 if ( $res===false ) die(__LINE__ . " - ".print_r($DBH->errorInfo(), true));
 $res = $DBH->exec($tables['Login_attempts'][$config['DB']['type']]);                // execute query for Login_attempts
 if ( $res===false ) die(__LINE__ . " - ".print_r($DBH->errorInfo(), true));
+$res = $DBH->exec($tables['Sessions'][$config['DB']['type']]);                      // execute query for Sessions
+if ( $res===false ) die(__LINE__ . " - ".print_r($DBH->errorInfo(), true));
 
 $retval['success'] = true;
 json_output_and_die($retval);