+++ inst/tables_specification.php
@@ -171,5 +171,19 @@ $tables = array(
     'sqlite' => "CREATE TABLE IF NOT EXISTS `".TBL_LOGIN_ATTEMPTS."` (
                           `Login_attempts_User_id`   INT NOT NULL     DEFAULT '0',
                           `Login_attempts_time`      INT NOT NULL     DEFAULT '0');"
+  ),
+  'Sessions' => array(
+    'mysql'  => "CREATE TABLE IF NOT EXISTS `".TBL_SESSIONS."` (
+                          `Session_id`               CHAR(128) NOT NULL DEFAULT '',
+                          `Session_set_time`         INT       NOT NULL DEFAULT '0',
+                          `Session_data`             TEXT      NOT NULL DEFAULT '',
+                          `Session_key`              CHAR(128) NOT NULL DEFAULT '',
+                          PRIMARY KEY (`Session_id`)
+                        ) ENGINE=MyISAM DEFAULT CHARSET=latin1 AUTO_INCREMENT=1;",
+    'sqlite' => "CREATE TABLE IF NOT EXISTS `".TBL_SESSIONS."` (
+                          `Session_id`               CHAR(128) NOT NULL DEFAULT '',
+                          `Session_set_time`         INT       NOT NULL DEFAULT '0',
+                          `Session_data`             TEXT      NOT NULL DEFAULT '',
+                          `Session_key`              CHAR(128) NOT NULL DEFAULT '');"
   )
 );